backend/requirements.txt
pymysql
drf_yasg==1.20.0
Django==3.1.2
djangorestframework==3.12.2
mysqlclient==2.0.3

requests~=2.23.0
pytest~=6.2.2
apscheduler~=3.7.0
pytz~=2021.1
itsdangerous~=1.1.0
flask~=1.1.2
jieba~=0.42.1